+++ desktop/src/main/kotlin/vn/loi/learning/desktop/ui/theme/LEIconsTokens.kt
@@ -37,6 +37,8 @@ data class LEIconsTokens(
     val Settings: ImageVector = Icons.Default.Settings,
     val Learning: ImageVector = Icons.Default.School,
     val Scheduler: ImageVector = Icons.Default.Schedule,
+    val Timer: ImageVector = Icons.Default.Timer,
+    val Info: ImageVector = Icons.Default.Info,
     val StatisticsTotal: ImageVector = Icons.Default.Layers,
     val StatisticsNew: ImageVector = Icons.Default.AddCircle,
     val StatisticsReview: ImageVector = Icons.Default.Refresh,